+++ insane-nights-days-mobile/components/nox/index.js
@@ -3,3 +3,4 @@ export { default as NoxButton } from './NoxButton';
 export { default as NoxInput } from './NoxInput';
 export { default as NoxCard } from './NoxCard';
 export { default as NoxScreenHeader } from './NoxScreenHeader';
+export { default as NoxSearchBar } from './NoxSearchBar';